requirements.txt
# Deep Learning Framework (updated for Python 3.14 compatibility)
torch>=2.9.0
torchvision>=0.20.0
pillow>=10.0.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# ML Experiment Tracking
mlflow>=2.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
opencv-python>=4.8.0

# Data Versioning
dvc>=3.30.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Monitoring & Logging
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
